+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## Changes
 - `Alphabet` now requires the size of the alphabet to be a power of 2 (breaking change)
+- Rename `gen_id` to just `gen` (breaking change)